+++ apps/docs/src/app/documentation/core-helpers/toolbar/toolbar.component.ts
@@ -22,7 +22,12 @@ export class ToolbarDocsComponent implements OnInit {
 
     library: string;
 
-    version: string = environment.version;
+    version = {
+        id: environment.version,
+        url: ''
+    };
+
+    versions: any[];
 
     items: ShellbarMenuItem[] = [
         {
@@ -69,9 +74,29 @@ export class ToolbarDocsComponent implements OnInit {
 
     ngOnInit(): void {
         this.cssUrl = this.sanitizer.bypassSecurityTrustResourceUrl('assets/sap_fiori_3.css');
+
+        this.versions = [
+            {id: '0.21.0', url: 'https://5f355f63718e9200075585e1--fundamental-ngx.netlify.app/'},
+            {id: '0.20.0', url: 'https://5f0630964a7a370007f93dc4--fundamental-ngx.netlify.app/'},
+            {id: '0.19.0', url: 'https://5ef288ca158ebd0008946f4d--fundamental-ngx.netlify.app/'},
+            {id: '0.18.0', url: 'https://5ec04b7f46b9bd000648f8ec--fundamental-ngx.netlify.app/'},
+            {id: '0.17.0', url: 'https://5e9a135cc7c8e90006047bdf--fundamental-ngx.netlify.app/'},
+            {id: '0.16.0', url: 'https://5e97032838070600063141e4--fundamental-ngx.netlify.app/'},
+            {id: '0.15.0', url: 'https://5e5fe7518009de0008f41fff--fundamental-ngx.netlify.app/'},
+            {id: '0.14.0', url: 'https://5e4f1d0714bc4c000ae3282d--fundamental-ngx.netlify.app/'},
+            {id: '0.13.0', url: 'https://5e25d4d1837fae0009b5c4fa--fundamental-ngx.netlify.app/'},
+            {id: '0.12.0', url: 'https://5db1dc978d2a340009a82d64--fundamental-ngx.netlify.app/'},
+            {id: '0.11.0', url: 'https://5d8a3409acaf8d00070ccd64--fundamental-ngx.netlify.app/'}
+        ];
+
+        this.versions.unshift(this.version);
     }
 
     selectTheme(selectedTheme: string): void {
         this.cssUrl = this.sanitizer.bypassSecurityTrustResourceUrl('assets/' + selectedTheme + '.css');
+    }
+
+    selectVersion(version: any): void {
+        window.open(version.url, '_blank');
     }
 }