+++ apps/docs/src/app/platform/component-docs/platform-dynamic-page/platform-dynamic-page-examples/platform-dynamic-page-example.component.ts
@@ -1,0 +1,33 @@
+import { ChangeDetectionStrategy, Component, ElementRef, ViewChild } from '@angular/core';
+import { DynamicPageCollapseChangeEvent } from '@fundamental-ngx/platform';
+
+@Component({
+    selector: 'fdp-dynamic-page-example',
+    templateUrl: './platform-dynamic-page-example.component.html',
+    styleUrls: ['./platform-dynamic-page-example.component.scss'],
+    changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class PlatformDynamicPageExampleComponent {
+    @ViewChild('overlay')
+    overlay: ElementRef<HTMLElement>;
+
+    fullscreen = false;
+
+    onCollapseChange(event: DynamicPageCollapseChangeEvent): void {
+        console.log('collapse changed');
+    }
+
+    resizeClicked(event: Event): void {
+        event.stopPropagation();
+    }
+
+    openPage(): void {
+        this.fullscreen = true;
+        this.overlay.nativeElement.style.width = '100%';
+    }
+    closePage(event: Event): void {
+        event.stopPropagation();
+        this.fullscreen = false;
+        this.overlay.nativeElement.style.width = '0%';
+    }
+}