+++ apps/docs/src/app/platform/component-docs/platform-object-list-item/platform-object-list-item-examples/platform-object-list-item-example.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fdp-object-list-item-example',
+    templateUrl: './platform-object-list-item-example.component.html'
+})
+export class PlatformObjectListItemExampleComponent { }