+++ apps/docs/src/app/platform/component-docs/platform-table/platform-table-examples/platform-table-multiple-row-selection-example.component.ts
@@ -1,0 +1,15 @@
+import { Component } from '@angular/core';
+
+import { ITEMS } from '../platform-table-docs.component';
+
+@Component({
+    selector: 'fdp-table-multiple-row-selection-example',
+    templateUrl: './platform-table-multiple-row-selection-example.component.html'
+})
+export class PlatformTableMultipleRowSelectionExampleComponent {
+    source: any[] = ITEMS;
+
+    onRowSelectionChange(event): void {
+        console.log(event);
+    }
+}