+++ apps/docs/src/app/platform/platform-documentation.routes.ts
@@ -1,4 +1,5 @@
 import { Routes } from '@angular/router';
+
 import { PlatformDocumentationComponent } from './documentation/platform-documentation.component';
 import { PlatformHomeComponent } from './component-docs/platform-home/platform-home.component';
 import { NewComponentComponent } from './component-docs/new-component/new-component.component';
@@ -38,6 +39,13 @@ export const ROUTES: Routes = [
                     import(
                         './component-docs/platform-forms/platform-checkbox-group/platform-checkbox-group-docs.module'
                     ).then((m) => m.PlatformCheckboxGroupDocsModule)
+            },
+            {
+                path: 'dynamic-page',
+                loadChildren: () =>
+                    import('./component-docs/platform-dynamic-page/platform-dynamic-page.module').then(
+                        (m) => m.PlatformDynamicPageDocsModule
+                    )
             },
             {
                 path: 'link',
@@ -92,6 +100,13 @@ export const ROUTES: Routes = [
                     )
             },
             {
+                path: 'table',
+                loadChildren: () =>
+                    import('./component-docs/platform-table/platform-table.module').then(
+                        (m) => m.PlatformTableDocsModule
+                    )
+            },
+            {
                 path: 'textarea',
                 loadChildren: () =>
                     import('./component-docs/platform-forms/platform-textarea/platform-textarea-docs.module').then(
@@ -134,6 +149,19 @@ export const ROUTES: Routes = [
                     )
             },
             {
+                path: 'object-marker',
+                loadChildren: () =>
+                    import('./component-docs/platform-object-marker/platform-object-marker.module').then(
+                        (m) => m.PlatformObjectMarkerDocsModule
+                    )
+            },
+            {
+                path: 'object-attribute',
+                loadChildren: () =>
+                    import('./component-docs/platform-object-attribute/platform-object-attribute-docs.module').then((m) =>
+                        m.PlatformObjectAttributeDocsModule)
+            },
+            {
                 path: 'input-group',
                 loadChildren: () =>
                     import('./component-docs/platform-forms/input-group/platform-input-group-docs.module').then(
@@ -143,25 +171,48 @@ export const ROUTES: Routes = [
             {
                 path: 'combobox',
                 loadChildren: () =>
-                    import('./component-docs/platform-combobox/platform-combobox-docs.module')
-                        .then(m => m.PlatformComboboxDocsModule)
+                    import('./component-docs/platform-combobox/platform-combobox-docs.module').then(
+                        (m) => m.PlatformComboboxDocsModule
+                    )
             },
             {
                 path: 'list',
                 loadChildren: () =>
-                    import('./component-docs/platform-list/platform-list-docs.module').then((m) => m.PlatformListDocsModule)
+                    import('./component-docs/platform-list/platform-list-docs.module').then(
+                        (m) => m.PlatformListDocsModule
+                    )
             },
             {
                 path: 'standard-list-item',
                 loadChildren: () =>
-                    import('./component-docs/platform-standard-list-item/platform-standard-list-item-docs.module').then((m) =>
-                        m.PlatformStandardListItemDocsModule)
+                    import('./component-docs/platform-standard-list-item/platform-standard-list-item-docs.module').then(
+                        (m) => m.PlatformStandardListItemDocsModule
+                    )
             },
             {
                 path: 'thumbnail',
                 loadChildren: () =>
-                    import('./component-docs/platform-thumbnail/platform-thumbnail-docs.module').then((m) =>
-                        m.PlatformThumbnailDocsModule)
+                    import('./component-docs/platform-thumbnail/platform-thumbnail-docs.module').then(
+                        (m) => m.PlatformThumbnailDocsModule
+                    )
+            },
+            {
+                path: 'action-list-item',
+                loadChildren: () =>
+                    import('./component-docs/platform-action-list-item/platform-action-list-item-docs.module').then((m) =>
+                        m.PlatformActionListItemDocsModule)
+            },
+            {
+                path: 'display-list-item',
+                loadChildren: () =>
+                    import('./component-docs/platform-display-list-item/platform-display-list-item-docs.module').then((m) =>
+                        m.PlatformDisplayListItemDocsModule)
+            },
+            {
+                path: 'object-list-item',
+                loadChildren: () =>
+                    import('./component-docs/platform-object-list-item/platform-object-list-item-docs.module').then((m) =>
+                        m.PlatformObjectListItemDocsModule)
             }
         ]
     }