+++ libs/platform/src/index.ts
@@ -17,4 +17,7 @@ export * from './lib/components/object-status/public_api';
 export * from './lib/components/form/combobox';
 export * from './lib/components/list/public_api';
 export * from './lib/components/thumbnail/public_api';
-
+export * from './lib/components/object-marker/public_api';
+export * from './lib/components/object-attribute/public_api';
+export * from './lib/components/table/public_api';
+export * from './lib/components/dynamic-page/public_api';