+++ libs/platform/src/lib/components/object-attribute/public_api.ts
@@ -1,0 +1,2 @@
+export * from './object-attribute.component';
+export * from './object-attribute.module';